string interpolation for field values
with this change whenever you use the {xxx} notation within the string
value of a field being created in a `put` or `reduce` proc then flume
will attempt to replace the {xxx} with the value of the field in the
current point with the name xxx. If xxx is not found the string will
remain untouched.